elastic-opentelemetry-instrumentation-openai: add integrations tests and test on baseline (#28)
* elastic-opentelemetry-instrumentation-openai: test against baseline

1.26.0 is required for having usage data with streams but with some
small changes we can keep the baseline as old as 1.2.0.

* elastic-opentelemetry-instrumentation-openai: add openai integration tests

This adds a bunch of integration tests hitting openai. This requires
setting a variable `OPENAI_API_KEY` in github secrets. Tests themselves
are a copy of the recorded ones asserting values from the response data.

Integration tests are not run by default and can be called with `pytest
--integration-tests`